refactored client code
