+++ src/client/src/App.tsx
@@ -1,11 +1,11 @@
 import { useState, useEffect } from "react";
-import { Board } from "./components/Board";
-import "./App.css";
-import { Login } from "./components/Login";
+import { Board } from "./components/tsx/Board";
+import { Login } from "./components/tsx/Login";
 import { io, Socket } from "socket.io-client";
+import "./App.css";
 
-const SERVER_URL = "https://minesweeper-server-o2fa.onrender.com";
-// const SERVER_URL = "http://localhost:3000";
+// const SERVER_URL = "https://minesweeper-server-o2fa.onrender.com";
+const SERVER_URL = "http://localhost:3000";
 
 function App() {
   const [username, setUsername] = useState("");
@@ -25,7 +25,7 @@ function App() {
       <div className="app-container">
         <div className="game-container">
           <h1 className="game-title">Co-op Minesweeper</h1>
-          <Board username={username} socket={socket} />
+          <Board uuid={username} socket={socket} />
         </div>
       </div>
     );
