+++ src/server/src/game/GameState.ts
@@ -1,30 +1,10 @@
-import { BoardConfig, Coord, Cell } from "../types/gameTypes";
-import { User, Dictionary } from "../types/serverTypes";
-import { Socket } from "socket.io";
-
-enum GameStatus {
-  PLAYING,
-  WON,
-  LOST,
-}
-
-type Direction = {
-  dx: number;
-  dy: number;
-};
-
-type GameSettings = {
-  width: number;
-  height: number;
-  mines: number;
-};
-
-export type LobbyState = {
-  board: GameState;
-  config: GameSettings;
-  connections: Dictionary<Socket>;
-  users: Dictionary<User>;
-};
+import {
+  BoardConfig,
+  Coord,
+  Cell,
+  GameStatus,
+  Direction,
+} from "../types/gameTypes";
 
 export class GameState {
   private readonly board: Cell[][];
@@ -67,7 +47,7 @@ export class GameState {
             isRevealed: false,
             adjMines: 0,
             isFlagged: false,
-          })),
+          }))
       );
   }
 
@@ -178,7 +158,7 @@ export class GameState {
 
   private checkWin(): boolean {
     return this.board.every((row) =>
-      row.every((cell) => cell.isMine || cell.isRevealed),
+      row.every((cell) => cell.isMine || cell.isRevealed)
     );
   }
 
@@ -314,7 +294,7 @@ export class GameState {
           isFlagged: cell.isFlagged,
           adjMines: cell.isRevealed ? cell.adjMines : null,
           isMine: this.status !== GameStatus.PLAYING ? cell.isMine : null,
-        })),
+        }))
       ),
       status: this.status,
       flagsLeft: this.flagsLeft,
