+++ src/server/src/types/serverTypes.ts
@@ -1,3 +1,7 @@
+import { GameState } from "../game/GameState";
+import { GameSettings } from "./gameTypes";
+import { Socket } from "socket.io";
+
 export interface User {
   uuid: string;
   state: {
@@ -9,3 +13,10 @@ export interface User {
 export interface Dictionary<T> {
   [key: string]: T;
 }
+
+export type LobbyState = {
+  board: GameState;
+  config: GameSettings;
+  connections: Dictionary<Socket>;
+  users: Dictionary<User>;
+};
